+++ gui/extension/CHANGELOG.md
@@ -18,6 +18,7 @@
 - BUG#38250829 SDK documentation does not list requirement for Python version
 - BUG#38262549 Unexpected compiler error for "delete" command in the TypeScript SDK
 - BUG#38260801 Unexpected error when generating the SDK for the default language
+- BUG#38257852 MySQL REST Service generates error code with a function without parameters
 -
 
 ## Changes in 1.19.15+9.4.0