+++ mrs_plugin/db_schema/mysql_rest_service_metadata.msm.project/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Database Schema Change Log for `mysql_rest_service_metadata`
 
+## 4.1.4
+
+- Update the `sdk_service_data` PROCEDURE to aggregate objects ordered by position
+
 ## 4.1.3
 
 - Fixed wrong time unit when compressing router_status log entries